migrate from inmem to jpa, refactor

+++ src/main/java/ua/rd/pizzaservice/repository/PizzaRepository.java
@@ -2,11 +2,9 @@ package ua.rd.pizzaservice.repository;
 
 import ua.rd.pizzaservice.domain.Pizza;
 
-/**
- * Created by Kostiantyn_Minkov on 10/4/2016.
- */
+
 public interface PizzaRepository {
 
-    Pizza findById(Integer id);
+    Pizza findById(Long id);
     Pizza save(Pizza pizza);
 }
